feat(service): 统一用户ID和服务ID参数类型为Long，提升数据一致性
- 将ClusterRoleUserService和DocService中的用户ID和服务ID参数类型从Integer更改为Long，以确保与数据模型的一致性。
- 更新相关实现类中的方法，确保系统在处理用户ID和服务ID时的稳定性与兼容性。

这些改动增强了服务层的健壮性与一致性，确保了整体设计的一致性。

+++ datasophon-service/src/main/java/com/datasophon/api/service/impl/ClusterAlertQuotaServiceImpl.java
@@ -93,7 +93,7 @@ public class ClusterAlertQuotaServiceImpl
                 .collect(Collectors.toSet());
 
         // 查询通知组 - ✅ NoticeGroupService已完成改造，启用代码
-        Set<Integer> noticeGroupIdList = alertQuotaList.stream()
+        Set<Long> noticeGroupIdList = alertQuotaList.stream()
                 .map(ClusterAlertQuotaEntity::getNoticeGroupId)
                 .collect(Collectors.toSet());
